Port code to be in datafusion
