+++ datafusion/sqllogictest/src/test_context.rs
@@ -15,17 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow_schema::TimeUnit;
 use async_trait::async_trait;
 use datafusion::execution::context::SessionState;
+use datafusion::logical_expr::Expr;
 use datafusion::physical_plan::ExecutionPlan;
+use datafusion::prelude::SessionConfig;
 use datafusion::{
     arrow::{
         array::{
             BinaryArray, Float64Array, Int32Array, LargeBinaryArray, LargeStringArray,
             StringArray, TimestampNanosecondArray,
         },
-        datatypes::{DataType, Field, Schema, SchemaRef},
+        datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit},
         record_batch::RecordBatch,
     },
     catalog::{schema::MemorySchemaProvider, CatalogProvider, MemoryCatalogProvider},
@@ -33,12 +34,102 @@ use datafusion::{
     prelude::{CsvReadOptions, SessionContext},
 };
 use datafusion_common::DataFusionError;
-use datafusion_expr::Expr;
+use log::info;
 use std::fs::File;
 use std::io::Write;
+use std::path::Path;
 use std::sync::Arc;
+use tempfile::TempDir;
 
-use crate::TestContext;
+/// Context for running tests
+pub struct TestContext {
+    /// Context for running queries
+    ctx: SessionContext,
+    /// Temporary directory created and cleared at the end of the test
+    test_dir: Option<TempDir>,
+}
+
+impl TestContext {
+    pub fn new(ctx: SessionContext) -> Self {
+        Self {
+            ctx,
+            test_dir: None,
+        }
+    }
+
+    /// Create a SessionContext, configured for the specific test, if
+    /// possible.
+    ///
+    /// If `None` is returned (e.g. because some needed feature is not
+    /// enabled), the file should be skipped
+    pub async fn try_new_for_test_file(relative_path: &Path) -> Option<Self> {
+        let config = SessionConfig::new()
+            // hardcode target partitions so plans are deterministic
+            .with_target_partitions(4);
+
+        let test_ctx = TestContext::new(SessionContext::with_config(config));
+
+        let file_name = relative_path.file_name().unwrap().to_str().unwrap();
+        match file_name {
+            "scalar.slt" => {
+                info!("Registering scalar tables");
+                register_scalar_tables(test_ctx.session_ctx()).await;
+            }
+            "information_schema_table_types.slt" => {
+                info!("Registering local temporary table");
+                register_temp_table(test_ctx.session_ctx()).await;
+            }
+            "information_schema_columns.slt" => {
+                info!("Registering table with many types");
+                register_table_with_many_types(test_ctx.session_ctx()).await;
+            }
+            "avro.slt" => {
+                #[cfg(feature = "avro")]
+                {
+                    let mut test_ctx = test_ctx;
+                    info!("Registering avro tables");
+                    register_avro_tables(&mut test_ctx).await;
+                    return Some(test_ctx);
+                }
+                #[cfg(not(feature = "avro"))]
+                {
+                    info!("Skipping {file_name} because avro feature is not enabled");
+                    return None;
+                }
+            }
+            "joins.slt" => {
+                info!("Registering partition table tables");
+
+                let mut test_ctx = test_ctx;
+                register_partition_table(&mut test_ctx).await;
+                return Some(test_ctx);
+            }
+            _ => {
+                info!("Using default SessionContext");
+            }
+        };
+        Some(test_ctx)
+    }
+
+    /// Enables the test directory feature. If not enabled,
+    /// calling `testdir_path` will result in a panic.
+    pub fn enable_testdir(&mut self) {
+        if self.test_dir.is_none() {
+            self.test_dir = Some(TempDir::new().expect("failed to create testdir"));
+        }
+    }
+
+    /// Returns the path to the test directory. Panics if the test
+    /// directory feature is not enabled via `enable_testdir`.
+    pub fn testdir_path(&self) -> &Path {
+        self.test_dir.as_ref().expect("testdir not enabled").path()
+    }
+
+    /// Returns a reference to the internal SessionContext
+    pub fn session_ctx(&self) -> &SessionContext {
+        &self.ctx
+    }
+}
 
 #[cfg(feature = "avro")]
 pub async fn register_avro_tables(ctx: &mut crate::TestContext) {
